new szenvedés

+++ src/main/java/proghf/controller/TableCollectionElementController.java
@@ -2,6 +2,8 @@ package proghf.controller;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import proghf.TableManager;
 import proghf.view.TableCollectionElementView;
@@ -24,8 +26,15 @@ public class TableCollectionElementController {
 
     @FXML
     public void onDeletePressed(ActionEvent actionEvent) {
-        var table = tableCollectionElementView.getTable();
-        table.getParent().deleteTable(table);
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Törlés megerősítése");
+        alert.setHeaderText("Biztosan törli a táblát?");
+        alert.setContentText("A tábla törlésével elvesznek az adatok.");
+        var result = alert.showAndWait();
+        if (result.get() == ButtonType.OK) {
+            var table = tableCollectionElementView.getTable();
+            table.getParent().deleteTable(table);
+        }
     }
 
     @FXML
